Add notes on how to write an adapter.
Add a more detailed description of how the gmail adapter works with
the juttle compiler and runtime to select messages and emit them as
points. This will hopefully be useful for others that want to write
adapters.

It has notes on:

* The concepts of live vs historical programs.
* The concepts of fields vs free text searches.
* How the read proc controls the timerange of searches and has a filter expression.
* How an adapter is loaded as a javascript module and what the main function should return.
* How the read/write operators map to Juttle.proc.base classes.
* How to emit errors for invalid options or filter expressions.
* How to work with the filter_ast object.
* How to work with the scheduler for live programs.
* How config in the .juttle/config.js file maps to the object passed to the main function.

At this point this is good enough as a first draft. I'll next work on
changing the adapter itself to support filter expressions and live
programs in the way these notes hint at.