Get live programs working.
The adapter now honors a -to that may be in the future. When -to is in
the future, it polls until the current time is greater than -to.

This required adding an intermediate function
get_messages_for_timerange that handled the details of creating the
gmail search expression, sorting the results, etc, as well as using
the scheduler to schedule a future call to get_messages_for_timerange().

Also fix a bug with -to handling--when rounded to a date for the gmail
search string, it should always be for a date after the current date.