Clean up -from/-to/-last handling.
Make sure -from/-to/-last arguments are handled in the way that
https://github.com/juttle/juttle/wiki/Time-Range-Semantics
describes.

Add tests for error cases.

One case (no -from/-to/-last) is waiting on a juttle release that
includes the new error code for that particular error. Until then,
throw a generic error and skip that testcase.

Also add a reference to the wiki page from the adapter implementation
notes page.