Address feedback from PR:
- Don't need both =~ and ~, ~ is turned into =~.
- Initialize filter_search_expr to undefined instead of ''.

The big change is to add a comprehensive set of unit tests for filter
expressions, in filter-gmail-compiler.spec.js. The unit tests handle
all the invalid and valid cases I could think of.

In order to easily run the GmailFilterCompiler as a standalone object,
I also stopped passing the program to the GmailFilterCompiler object
and switched to using JuttleErrors directly to raise
RT-ADAPTER-UNSUPPORTED-FILTER errors.