Add write proc that sends emails.
Add a new proc in write.js that handles writing emails. It uses the
gmail API to send one or more emails containing the points as
stringified JSON.

Expose the write proc as a part of the module.

Modify the create_oauth_token.js script to create tokens that also
have the ability to send emails as well as read them.

Also add basic unit tests that send emails to whatever account is
specified by the JUTTLE_GMAIL_CONFIG environment variable.