Reorganize and add docs on write adapter.
Split adapter_impl_notes into separate docs for general adapter
configuration, details on the read half of the adapter (in read.md),
and details on the write half of the adapter (in write.md).

The write half has notes on what it receives (json points) how it
interacts with the program for orderly shutdown, and a description of
the main methods.

Also add notes on write proc options to the main README.md file.