Update for juttle 0.5.0.
Copy dot files from juttle-adapter-template to switch to gulp and eslint.

Install juttle ^0.5.0 in travis instead of the latest juttle.

Add a juttleAdapterAPI element to package.json with version 0.5.0.

Also update defaultTimeRange to defaultTimeOptions.

Make everything ES6 friendly with classes, fat arrow for loops, no
more self, let instead of var, for (item of array) instead of forEach,
etc.

Also fix everything found by eslint, mostly use of double quotes.

In FilterGmailCompiler, make the recommendations suggested in
https://github.com/juttle/juttle/wiki/Porting-Adapters-to-the-0.5.x-API:

 - no visitExpressionFilterTerm
 - rely on visitFulltextFilterTerm instead of visitSimpleFilterTerm
 - use visitField instead of visitUnaryExpression + '*'.
 - use featureNotSupported to return errors instead of throwing our own.

In the filter tests, this means creating a FilterSimplifier and
running it on the ast.

In the main index.js, validate that the config has the keys we want.

In the adapter classes, switch to a class deriving from
AdapterRead/AdapterWrite and use JuttleAdapterAPI to access
everything. Use the built-in code for validating options instead of
doing our own checks. Get rid of interactions with the scheduler.

In write, use the sql adapter's trick of keeping an ongoing promise
that is appended to as more write()s are called, and is returned in
eof().